@@ -18,65 +18,65 @@
         "compodoc": "./node_modules/.bin/compodoc -d ./docs/ -p ./tsconfig.json"
     },
     "dependencies": {
-        "@angular/animations": "5.2.11",
-        "@angular/common": "5.2.11",
-        "@angular/compiler": "5.2.11",
-        "@angular/compiler-cli": "5.2.11",
-        "@angular/core": "5.2.11",
-        "@angular/forms": "5.2.11",
-        "@angular/http": "5.2.11",
-        "@angular/platform-browser": "5.2.11",
-        "@angular/platform-browser-dynamic": "5.2.11",
-        "@ionic-native/core": "^4.8.0",
-        "@ionic-native/file": "^4.8.0",
-        "@ionic-native/keyboard": "4.8.0",
-        "@ionic-native/push": "4.8.0",
-        "@ionic-native/splash-screen": "4.8.0",
-        "@ionic-native/sqlite": "4.8.0",
-        "@ionic-native/sqlite-porter": "4.8.0",
-        "@ionic-native/status-bar": "4.8.0",
+        "@angular/animations": "7.2.9",
+        "@angular/common": "7.2.9",
+        "@angular/compiler": "7.2.9",
+        "@angular/compiler-cli": "7.2.9",
+        "@angular/core": "7.2.9",
+        "@angular/forms": "7.2.9",
+        "@angular/http": "7.2.9",
+        "@angular/platform-browser": "7.2.9",
+        "@angular/platform-browser-dynamic": "7.2.9",
+        "@ionic-native/core": "^5.2.0",
+        "@ionic-native/file": "^5.2.0",
+        "@ionic-native/keyboard": "5.2.0",
+        "@ionic-native/push": "5.2.0",
+        "@ionic-native/splash-screen": "5.2.0",
+        "@ionic-native/sqlite": "5.2.0",
+        "@ionic-native/sqlite-porter": "5.2.0",
+        "@ionic-native/status-bar": "5.2.0",
         "@ionic/cli-plugin-cordova": "1.6.2",
-        "@ionic/storage": "2.1.3",
-        "@ngx-translate/core": "^9.1.1",
-        "@ngx-translate/http-loader": "^2.0.1",
+        "@ionic/storage": "2.2.0",
+        "@ngx-translate/core": "^11.0.1",
+        "@ngx-translate/http-loader": "^4.0.0",
         "add": "^2.0.6",
         "build": "^0.1.4",
-        "cordova-android": "^7.1.4",
-        "cordova-browser": "^5.0.4",
-        "cordova-ios": "^4.5.4",
-        "cordova-osx": "~4.0.1",
+        "cordova-android": "^8.0.0",
+        "cordova-browser": "^6.0.0",
+        "cordova-ios": "^5.0.0",
+        "cordova-osx": "~5.0.0",
         "cordova-plugin-console": "^1.1.0",
-        "cordova-plugin-device": "^1.1.7",
-        "cordova-plugin-splashscreen": "^4.1.0",
+        "cordova-plugin-device": "^2.0.2",
+        "cordova-plugin-splashscreen": "^5.0.2",
         "cordova-plugin-statusbar": "^2.4.2",
         "cordova-plugin-whitelist": "^1.3.3",
-        "cordova-sqlite-storage": "^2.6.0",
+        "cordova-sqlite-storage": "^3.2.0",
         "device": "^0.3.9",
-        "express": "^4.15.4",
-        "firebase": "^5.0.2",
-        "ionic-angular": "3.9.2",
+        "express": "^4.16.4",
+        "firebase": "5.9.0",
+        "ionic-angular": "3.9.4",
         "ionic-plugin-keyboard": "^2.2.1",
-        "ionicons": "3.0.0",
+        "ionicons": "4.5.5",
         "material-design-icons": "^3.0.1",
-        "moment": "^2.20.1",
-        "npm": "^4.6.1",
+        "moment": "^2.24.0",
+        "npm": "^6.9.0",
         "phonegap-plugin-push": "^2.2.3",
-        "promise-polyfill": "6.1.0",
-        "rxjs": "5.5.11",
+        "promise-polyfill": "8.1.0",
+        "rxjs": "6.4.0",
+        "rxjs-compat": "^6.4.0",
         "sw-toolbox": "3.6.0",
         "uk.co.workingedge.cordova.plugin.sqliteporter": "^1.1.0",
-        "zone.js": "0.8.26"
+        "zone.js": "0.8.29"
     },
     "resolutions": {
         "grpc": "^1.12.0"
     },
     "devDependencies": {
-        "@angular/cli": "^1.7.4",
+        "@angular/cli": "^7.3.6",
         "@angular/tsc-wrapped": "^4.4.6",
-        "@ionic/app-scripts": "^3.2.1",
-        "@ionic/cli-plugin-proxy": "1.5.8",
-        "ionic": "3.20.0",
-        "typescript": "^2.6.2"
+        "@ionic/app-scripts": "^3.2.3",
+        "@ionic/cli-plugin-proxy": "^1.5.9",
+        "typescript": "3.3.3333"
     },
     "cordovaPlugins": [
         "cordova-plugin-whitelist",
