@@ -2,9 +2,9 @@
 // These scripts are made available when the app is served or deployed on Firebase Hosting
 // If you do not serve/host your project using Firebase Hosting see https://firebase.google.com/docs/web/setup
 
-importScripts('https://www.gstatic.com/firebasejs/4.2.0/firebase-app.js');
-importScripts('https://www.gstatic.com/firebasejs/4.2.0/firebase-messaging.js');
-importScripts('https://www.gstatic.com/firebasejs/4.2.0/firebase.js');
+importScripts('https://www.gstatic.com/firebasejs/5.8.6/firebase-app.js');
+importScripts('https://www.gstatic.com/firebasejs/5.8.6/firebase-messaging.js');
+importScripts('https://www.gstatic.com/firebasejs/5.8.6/firebase.js');
 
 
 
